+++ publish.sh
@@ -4,4 +4,3 @@ uvx hatch clean
 gitnextver .
 uvx hatch build
 uv publish
-c